Update to the docs app
